+++ packages/dash-core-components/babel.config.js
@@ -4,9 +4,7 @@ const presets = [
 ];
 
 const plugins = [
-    '@babel/plugin-syntax-dynamic-import',
-    '@babel/plugin-transform-async-to-generator',
-    '@babel/plugin-transform-runtime'
+    '@babel/plugin-syntax-dynamic-import'
 ];
 
 // eslint-disable-next-line no-process-env