+++ packages/dash-core-components/src/utils/MarkdownHighlighter.js
@@ -1,10 +1,12 @@
 import lazyhljs from './LazyLoader/hljs';
 
 const MarkdownHighlighter = {
-    loadhljs: async function() {
-        this.hljs = await lazyhljs();
-        this.hljsResolve();
-        this.isReady = true;
+    loadhljs: function() {
+        return lazyhljs().then(hljs => {
+            this.hljs = hljs;
+            this.hljsResolve();
+            this.isReady = true;
+        });
     },
 };