+++ packages/dash-core-components/webpack.config.js
@@ -60,9 +60,24 @@ module.exports = (env, argv) => {
                     test: /\.jsx?$/,
                     exclude: /node_modules/,
                     use: {
-                        loader: 'babel-loader',
-                    },
+                        loader: 'babel-loader'
+                    }
                 },
+                {
+                    test: /\.jsx?$/,
+                    include: /node_modules\/(react-jsx-parser\/)/,
+                    use: {
+                        loader: 'babel-loader',
+                        options: {
+                            babelrc: false,
+                            configFile: false,
+                            presets: [
+                                '@babel/preset-env'
+                            ]
+                        }
+                    }
+                },
+
                 {
                     test: /\.css$/,
                     use: [
@@ -105,6 +120,12 @@ module.exports = (env, argv) => {
                         name(module, chunks, cacheGroupKey) {
                             return `${cacheGroupKey}~${chunks[0].name}`;
                         }
+                    },
+                    shared: {
+                        chunks: 'all',
+                        minSize: 0,
+                        minChunks: 2,
+                        name: 'dash_core_components-shared'
                     }
                 }
             }